Updated to 1.13, compiles, can somewhat run till init before hopelessly collapsing in on itself Some things (besides the fact that it doesn't load at all) don't work (as intended) as Forge isn't ready yet

+++ src/main/java/elec332/core/inventory/widget/WidgetButtonArrow.java
@@ -21,7 +21,7 @@ public class WidgetButtonArrow extends WidgetButton {
     }
 
     @Override
-    public void draw(Window gui, int guiX, int guiY, int mouseX, int mouseY) {
+    public void draw(Window gui, int guiX, int guiY, double mouseX, double mouseY) {
         bindTexture(new ResourceLocation("eleccore", "buttons.png"));
         int u = this.u;
         int v = this.v;
